Add Error Handling and Swagger and Refactoring Code

+++ API/Controllers/CategoryController.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using API.DTOs.CategoryDtos;
@@ -6,7 +7,7 @@ using API.DTOs.SubCategoryDtos;
 using AutoMapper;
 using Core.Entities;
 using Core.Interfaces;
-using Interfaces.Core;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers
@@ -17,22 +18,24 @@ namespace API.Controllers
     {
         private readonly IGenericRepository<Category, int> _categoryRepo;
         private readonly IGenericRepository<SubCategory, int> _subCategoryRepo;
+        private readonly IUserRepository _userRepo;
         private readonly IMapper _mapper;
 
         public CategoryController(IGenericRepository<Category, int> categoryRepo,
-            IGenericRepository<SubCategory, int> subCategoryRepo , IMapper mapper)
+            IGenericRepository<SubCategory, int> subCategoryRepo, IUserRepository userRepo, IMapper mapper)
         {
             _categoryRepo = categoryRepo;
             _subCategoryRepo = subCategoryRepo;
+            _userRepo = userRepo;
             _mapper = mapper;
         }
 
         [HttpGet]
         public async Task<IReadOnlyList<CategoryToReturnDto>> GetAll()
         {
-             var categories = await _categoryRepo.GetAll();
-             var data = _mapper.Map<IReadOnlyList<Category>, IReadOnlyList<CategoryToReturnDto>>(categories);
-             return data.ToList();
+            var categories = await _categoryRepo.GetAll();
+            var data = _mapper.Map<IReadOnlyList<Category>, IReadOnlyList<CategoryToReturnDto>>(categories);
+            return data.ToList();
         }
 
         [HttpGet("{id}")]
@@ -41,12 +44,16 @@ namespace API.Controllers
             var category = await _categoryRepo.GetById(id);
             var data = _mapper.Map<Category, CategoryToReturnDto>(category);
             if (category != null) return Ok(data);
-            return BadRequest("This Category is not exist.");
+            throw new Exception("Error happen when get Category, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Add(Category category)
         {
+            var user = await _userRepo.GetUserByUserClaims(HttpContext.User);
+            if (user == null) return Unauthorized("User is Unauthorized");
+
             if (await _categoryRepo.GetById(category.Id) != null)
                 return BadRequest("This category is exist.");
 
@@ -57,31 +64,39 @@ namespace API.Controllers
 
             if (await _categoryRepo.Add(categoryForAdd))
                 return Ok();
-            return BadRequest("Error happen when add Category.");
+            throw new Exception("Error happen when add Category, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> Update(int id, Category entity)
         {
+            var user = await _userRepo.GetUserByUserClaims(HttpContext.User);
+            if (user == null) return Unauthorized("User is Unauthorized");
+
             var category = await _categoryRepo.GetById(id);
             if (category == null)
                 return BadRequest("This category is not exist!");
             category.Name = entity.Name;
             if (await _categoryRepo.Update(category))
                 return Ok();
-            return BadRequest("Error happen when update Category.");
+            throw new Exception("Error happen when update Category, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> Delete(int id)
         {
+            var user = await _userRepo.GetUserByUserClaims(HttpContext.User);
+            if (user == null) return Unauthorized("User is Unauthorized");
+
             var category = await _categoryRepo.GetById(id);
             if (category == null)
                 return BadRequest("Please check Category exist!");
-            
+
             if (await _categoryRepo.Delete(id))
                 return Ok();
-            return BadRequest("Error happen when delete Category.");
+            throw new Exception("Error happen when delete Category, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
         }
 
         [HttpGet("GetSubCategories/{categoryId}")]
@@ -92,56 +107,70 @@ namespace API.Controllers
                 return null;
             var subCategories = await _subCategoryRepo.GetAll();
             subCategories = subCategories.Where(x => x.CategoryId == categoryId).ToList();
-            var subCategoriesToReturn = _mapper.Map<IReadOnlyList<SubCategory>,IReadOnlyList<SubCategoryToReturnDto>>(subCategories);
+            var subCategoriesToReturn = _mapper.Map<IReadOnlyList<SubCategory>, IReadOnlyList<SubCategoryToReturnDto>>(subCategories);
             return subCategoriesToReturn;
+            throw new Exception("Error happen when get SubCategories, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
+
         }
 
         [HttpPost("AddSubCategory")]
+        [Authorize]
         public async Task<IActionResult> AddSubCategory([FromBody] SubCategory entity)
         {
-            if (await _subCategoryRepo.GetById(entity.Id) !=null)
+            var user = await _userRepo.GetUserByUserClaims(HttpContext.User);
+            if (user == null) return Unauthorized("User is Unauthorized");
+
+            if (await _subCategoryRepo.GetById(entity.Id) != null)
                 return BadRequest("This SubCategory is exist");
-            
-            if (await _categoryRepo.GetById(entity.CategoryId) ==null)
+
+            if (await _categoryRepo.GetById(entity.CategoryId) == null)
                 return BadRequest(@"You Can't add SubCategory to not exist Category!");
-            
-            var subCategory =new SubCategory()
+
+            var subCategory = new SubCategory()
             {
                 Name = entity.Name,
                 CategoryId = entity.CategoryId
             };
-            
+
             if (await _subCategoryRepo.Add(subCategory))
                 return Ok();
-            return BadRequest("Error happen when add SubCategory.");
+            throw new Exception("Error happen when add Sub Category, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
         }
 
         [HttpPut("updateNameSubCategory/{id}")]
+        [Authorize]
         public async Task<IActionResult> UpdateNameSubCategory(int id, SubCategoryForUpdateDto entity)
         {
-            if (await _subCategoryRepo.GetById(id) ==null)
+            var user = await _userRepo.GetUserByUserClaims(HttpContext.User);
+            if (user == null) return Unauthorized("User is Unauthorized");
+
+            if (await _subCategoryRepo.GetById(id) == null)
                 return BadRequest("This SubCategory is not exist");
-            
+
             var subCategory = await _subCategoryRepo.GetById(id);
-            
-            if (await _categoryRepo.GetById(subCategory.CategoryId) ==null)
+
+            if (await _categoryRepo.GetById(subCategory.CategoryId) == null)
                 return BadRequest(@"You Can't update SubCategory to not exist Category!");
-            
+
             subCategory.Name = entity.Name;
             if (await _subCategoryRepo.Update(subCategory))
                 return Ok();
-            return BadRequest("Error happen when update name SubCategory.");
+            throw new Exception("Error happen when update Sub Category, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
         }
 
         [HttpDelete("DeleteSubCategory/{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteSubCategory(int id)
         {
+            var user = await _userRepo.GetUserByUserClaims(HttpContext.User);
+            if (user == null) return Unauthorized("User is Unauthorized");
+
             var subCategory = await _subCategoryRepo.GetById(id);
             if (subCategory == null)
                 return BadRequest("Please add SubCategory exist!");
             if (await _subCategoryRepo.Delete(subCategory.Id))
                 return Ok();
-            return BadRequest("Error happen when delete SubCategory.");
+            throw new Exception("Error happen when delete Sub Category, Ahmad Nour hate Exception ):,Exception hate Ahmad Nour ): please don't make any error, i see you *-*");
         }
     }
 }
